+++ CHANGELOG.md
@@ -17,6 +17,9 @@ notes below for additional usage guidelines.
 * Updated to pyjoulescope_driver 1.3.4 and sensor FPGA 1.0.4.
 * Fixed Memory Widget "Clear" not clearing when streaming is paused.
 * Hide status bar troubleshooting details by default.
+* Improved waveform time axis to display conventional time format.
+* Removed default quantization from time_map trel_offset.
+* Fixed x_range inaccuracy due to unit corruption (int->float) when pinned.
 
 
 ## 1.0.9